requirements.txt
openai
python-dotenv
flask
Flask-Cors